refactor: package code organization wave 19 (self-healing pure peels)
U5 Slice A: introduce packages/engine/src/self-healing/ and peel pure free
functions from self-healing.ts while keeping public imports stable via
re-exports.

- autoRecoverWorktreeSessionStartFailure → self-healing/auto-recover-worktree-session.ts
- archiveAsGhostBug → self-healing/archive-ghost-bug.ts
- step-progress predicates → self-healing/step-progress.ts
- ratchet self-healing.ts line-count baseline to post-peel live LOC
